random stuff

+++ nvim/lua/config/options.lua
@@ -35,7 +35,7 @@ vim.opt.clipboard:append("unnamedplus")
 -- Split window settings
 vim.opt.splitright = true
 vim.opt.splitbelow = true
-vim.o.winborder = 'double'
+vim.o.winborder = 'single'
 
 -- Make it so that semantic highlighting has priority over treesitter
 vim.highlight.priorities.semantic_tokens = 101
